+++ src/env.sh
@@ -11,6 +11,7 @@ print_var BASE_PROJECT_MANIFEST
 print_var BASE_PROJECT_VENV_DIR
 print_var BASE_OS
 print_var BASE_PLATFORM
+print_var BASE_HOST_ENV
 print_var BASE_HOST
 print_var BASE_DEMO_ENV
 print_var BASE_DEMO_ACTIVATED